refactor: restructure file paths for clarity
