+++ src/components/btc-mts/MTSSlider.tsx
