+++ src/components/btc/Slider.tsx
