+++ src/components/btc/use-pointer-interaction.ts
