+++ src/demos/BTCMTSColorPicker.tsx
@@ -2,7 +2,7 @@ import { root, runOnBackground, useCallback, useState } from '@lynx-js/react';
 import { AppLayout } from '@/App';
 import { sleep } from '@/utils/sleep';
 
-import { ColorPicker } from '@/components/btc-mts-colorpicker/MTSColorPicker';
+import { ColorPicker } from '@/components/btc-mts/MTSColorPicker';
 
 if (__BACKGROUND__) {
   setInterval(() => {
